Update SentryFlutterFFI to SentryFlutterPlugin
