+++ packages/flutter/ios/sentry_flutter/Sources/sentry_flutter_objc/SentryFlutterPlugin.h
@@ -3,7 +3,7 @@
 #if __has_include(<sentry_flutter/sentry_flutter-Swift.h>)
 #import <sentry_flutter/sentry_flutter-Swift.h>
 #else
-@interface SentryFlutterFFI : NSObject
+@interface SentryFlutterPlugin : NSObject
 + (nullable NSData *)loadContextsAsBytes;
 + (nullable NSData *)loadDebugImagesAsBytes:(NSSet<NSString *> *)instructionAddresses;
 @end
